+++ galleryfrontend/package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.27.2",
     "core-js": "^2.5.7",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
