+++ galleryfrontend/src/App.js
@@ -6,6 +6,7 @@ import './components/Login/login.css'
 import './components/Main/main.css'
 import { Routes, Route } from 'react-router-dom'
 import UpLoad from './components/PostReg/upload';
+import Test from './components/test';
 
 
 function App() {
@@ -15,6 +16,7 @@ function App() {
         <Route exact path="/" element={<Main></Main>}></Route>
         <Route path="/login" element={<Login></Login>}></Route>
         <Route path="/upload" element={<UpLoad></UpLoad>}></Route>
+        <Route path="/test" element={<Test></Test>}></Route>
       </Routes>
     </div>
   );
